login page bug fixed

+++ src/content/class/class.controller.ts
@@ -1,13 +1,29 @@
-import {Body, Controller, Delete, Get, Param, Post, Put} from '@nestjs/common';
+import {
+    Body,
+    Controller,
+    Delete,
+    ForbiddenException,
+    Get,
+    Param,
+    Post,
+    Put,
+    Render,
+    Req,
+    UnauthorizedException
+} from '@nestjs/common';
 import {ClassService} from "./class.service";
 import {ClassDto} from "../dto/class.dto";
+import {UserService} from "../../authentication/user/user.service";
 
 @Controller('api/class')
 export class ClassController {
-    constructor(protected readonly classService: ClassService) {}
+    constructor(protected readonly classService: ClassService,
+                protected readonly userService: UserService) {
+    }
 
     @Get('')
-    getAllClasses() {
+    @Render('class')
+    async getAllClasses(@Req() request) {
         return this.classService.getAllClasses();
     }
 
@@ -17,17 +33,30 @@ export class ClassController {
     }
 
     @Get(':id')
+    @Render('class/single.hbs')
     getClassById(@Param('id') id) {
         return this.classService.getClassById(id);
     }
 
     @Put(':id')
-    updateClassById(@Body() body, @Param('id') id) {
+    async updateClassById(@Req() request, @Body() body, @Param('id') id) {
+        const {auth} = request.cookies
+        const user = await this.userService.findById(auth)
+        if (!user)
+            throw new UnauthorizedException()
+        if (user.role !== 'ADMIN')
+            throw new ForbiddenException()
         return this.classService.updateClassById(body, id);
     }
 
     @Delete(':id')
-    deleteClassById(@Param('id') id) {
+    async deleteClassById(@Req() request, @Param('id') id) {
+        const {auth} = request.cookies
+        const user = await this.userService.findById(auth)
+        if (!user)
+            throw new UnauthorizedException()
+        if (user.role !== 'ADMIN')
+            throw new ForbiddenException()
         return this.classService.deleteClassById(id);
     }
 }
